move more files from app to feature folders
