+++ frontend/src/modules/app/index.ts
@@ -11,9 +11,6 @@
  *
  * Re-usable layout components.
  */
-export * from '../navigation/layouts';
 export * from './App';
-export * from './OverlayScreen';
 export * from './Service';
-export * from './ToolBar';
 export * from './UserActivity';
