+++ frontend/src/modules/navigation/layouts/ScreensaverRoute.tsx
@@ -12,10 +12,10 @@ import ClockIcon from '../../icons/ClockIcon';
 import { DASHBOARD_ROUTE, QUICKSTART_ROUTE } from '../constants';
 import Alarms from '../../alarms/Alarms';
 import UserActivity from '../../app/UserActivity';
-import OverlayScreen from '../../app/OverlayScreen';
+import OverlayScreen from '../../overlay/OverlayScreen';
 import { getAlarmNotifyStatus } from '../../../store/app/selectors';
-import PowerIndicator from '../../app/PowerIndicator';
-import HeaderClock from '../../app/HeaderClock';
+import PowerIndicator from '../../toolbar/PowerIndicator';
+import HeaderClock from '../../toolbar/HeaderClock';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
