+++ frontend/src/modules/overlay/index.ts
@@ -1,0 +1,1 @@
+export * from './OverlayScreen';
