+++ frontend/src/modules/toolbar/PowerIndicator.tsx
