+++ frontend/src/modules/toolbar/index.ts
@@ -1,0 +1,1 @@
+export * from './ToolBar';
